Add Boltzmann pdf sampling
The file boltzmann_network.py is our implementation of the hopfield
network with an energy dependent activation function. The function is
not complete and the precomputed exponentials have yet to be added to
the HopfieldNetwork class.

Lattice.py is a new class that has a useful method to get the values and
indices of nearest neighbours, which is used to calculate the change in
energy between states.

This is the same update I posted as in the repository by YaSleiMan, one of my
team members. We have switched control over the repository